close #6

+++ internal/pkg/models/url.go
@@ -10,7 +10,7 @@ type Url struct {
 	Model
 	OriginalUrl string `gorm:"type:varchar(255)" json:"url"`
 	ShortenUrl  string `gorm:"type:varchar(255)" json:"shorten_url"`
-	UserId      uint64 `gorm:"not null" json:"-" validation:"user_id"`
+	UserId      uint   `gorm:"not null" json:"-" validation:"user_id"`
 	User        User   `gorm:"foreignkey:UserId;constraint:onUpdate:CASCADE,onDelete:CASCADE" json:"user"`
 }
